weather/main/serializers.py
from main.models import City
from rest_framework import serializers


class CitySerializer(serializers.ModelSerializer):
    class Meta:
        model = City
        fields = ['name', 'requests_count']
